Memory management runtime overhaul
The runtime is now initialized from values put into static memory. This enables moving the initialization logic to the runtime itself, making it easier to add additional functionality that requires pre-initialized standard functions to be available right away, i.e. a GC implementation that needs to be able to call 'malloc' instead of 'mspace_malloc' etc.